backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
